event: Remove unnecessary casts
(cherry picked from commit 2854e81903118d3ac736f93151b5b11d7ee8b3b8)

+++ mcfgthread/event.c
@@ -26,7 +26,7 @@ _MCF_event_await_change_slow(_MCF_event* event, int undesired, const int64_t* ti
   try_wait_loop:
     _MCF_atomic_load_pptr_acq(&old, event);
     do {
-      if(old.__value != (uint8_t) undesired)
+      if(old.__value != undesired)
         return old.__value;
 
       new = old;
